Cart improvements

+++ src/App.js
@@ -18,6 +18,9 @@ async function fetchUser(setUser, featuredCourses) {
       let user = Object.values(users.data).filter(
         (user) => user.username === "Sandra"
       )[0];
+      user.wishlistCourses = user.wishlistCourses || [];
+      user.laterList = user.laterList || [];
+      let laterListSlugs = Object.entries(user.laterList);
       let wishlistSlugs = Object.entries(user.wishlistCourses);
       let myCoursesSlugs = Object.entries(user.myCourses);
       let cartSlugs = Object.entries(user.cart).filter(
@@ -32,9 +35,22 @@ async function fetchUser(setUser, featuredCourses) {
           };
           return course;
         }),
-        wishlistCourses: wishlistSlugs.map(([id, slug]) => {
+        wishlistCourses: wishlistSlugs.map(([id, wishlistItem]) => {
+          console.log(id, wishlistItem);
           let course = {
-            ...featuredCourses.filter((course) => course.slug === slug)[0],
+            ...featuredCourses.filter(
+              (course) => course.slug === wishlistItem.course
+            )[0],
+            id: id,
+          };
+          return course;
+        }),
+        laterListCourses: laterListSlugs.map(([id, laterListItem]) => {
+          console.log(id, laterListItem);
+          let course = {
+            ...featuredCourses.filter(
+              (course) => course.slug === laterListItem.course
+            )[0],
             id: id,
           };
           return course;
@@ -53,26 +69,73 @@ async function fetchUser(setUser, featuredCourses) {
   }
 }
 
-async function addCourseToWishlist(slug, user, setUser, featuredCourses) {
-  const duplicatedCourses = user.wishlistCourses.filter(
-    (course) => slug === course.slug
-  );
-  if (duplicatedCourses.length) {
-    try {
-      const userId = user.id;
-      const result = await axios.delete(
-        `Users/${userId}/wishlistCourses/${duplicatedCourses[0].id}.json`
-      );
-      if (result) {
-        return fetchUser(setUser, featuredCourses);
-      }
-    } catch (error) {}
-  }
-  try {
-    const propPrefix = "wishlistCourse";
+async function addCourseToWishlist(
+  slug,
+  user,
+  setUser,
+  featuredCourses,
+  deleteIfMatches = true
+) {
+  console.log(user);
+  const duplicatedCourses = user.wishlistCourses.filter((course) => {
+    return slug === course.slug;
+  });
+
+  // condición para borrar el elemento si esta duplicado y deleteIfMatches es verdadero
+  if (deleteIfMatches && duplicatedCourses.length) {
     const userId = user.id;
-    const result = await axios.patch(`/Users/${userId}/wishlistCourses.json`, {
-      [propPrefix + (user.wishlistCourses.length + 1)]: slug,
+    const result = await axios.delete(
+      `Users/${userId}/wishlistCourses/${duplicatedCourses[0].id}.json`
+    );
+    if (result) {
+      return fetchUser(setUser, featuredCourses);
+    }
+  }
+
+  // condición para saltarnos la ejecución si ya esta el curso en el wishlist
+  if (!deleteIfMatches && duplicatedCourses.length) return;
+
+  try {
+    const userId = user.id;
+    const result = await axios.post(`/Users/${userId}/wishlistCourses.json`, {
+      course: slug,
+    });
+    if (result) {
+      fetchUser(setUser, featuredCourses);
+    }
+  } catch (error) {}
+}
+
+async function addCourseToSaveforLaterList(
+  slug,
+  user,
+  setUser,
+  featuredCourses,
+  deleteIfMatches = true
+) {
+  console.log(user);
+  const duplicatedCourses = user.laterListCourses.filter((course) => {
+    return slug === course.slug;
+  });
+
+  // condición para borrar el elemento si esta duplicado y deleteIfMatches es verdadero
+  if (deleteIfMatches && duplicatedCourses.length) {
+    const userId = user.id;
+    const result = await axios.delete(
+      `Users/${userId}/laterList/${duplicatedCourses[0].id}.json`
+    );
+    if (result) {
+      return fetchUser(setUser, featuredCourses);
+    }
+  }
+
+  // condición para saltarnos la ejecución si ya esta el curso en el wishlist
+  if (!deleteIfMatches && duplicatedCourses.length) return;
+
+  try {
+    const userId = user.id;
+    const result = await axios.post(`/Users/${userId}/laterList.json`, {
+      course: slug,
     });
     if (result) {
       fetchUser(setUser, featuredCourses);
@@ -105,6 +168,7 @@ function generateRandomHash(propPrefix) {
 }
 
 async function addCourseToCart(slug, user, setUser, featuredCourses) {
+  console.log(user);
   const duplicatedCourses = user.cart.filter((course) => slug === course.slug);
   if (duplicatedCourses.length) {
     try {
@@ -203,12 +267,21 @@ function App() {
               <MyCourses
                 featuredCourses={featuredCourses}
                 user={user}
+                setUser={setUser}
                 activeSection={match.params.subsection}
+                addCourseToCart={addCourseToCart}
               />
             )}
           ></Route>
           <Route path="/cart">
-            <Cart user={user} setUser={setUser} featuredCourses={featuredCourses} addCourseToCart={addCourseToCart}/>
+            <Cart
+              user={user}
+              setUser={setUser}
+              featuredCourses={featuredCourses}
+              addCourseToCart={addCourseToCart}
+              addCourseToWishlist={addCourseToWishlist}
+              addCourseToSaveforLaterList={addCourseToSaveforLaterList}
+            />
           </Route>
           <Route
             path="/:genericSection"
